backend/pet_society/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.3.0
sqlparse==0.5.3
channels
cloudinary
daphne
dj-database-url
django-cleanup
django-cloudinary-storage
django-environ
django-htmx
psycopg2-binary
requests
shortuuid
